requirements.txt
psycopg2-binary>=2.9
paho-mqtt>=1.6
python-dotenv>=1.0
requests>=2.28
# Optional for ML
scikit-learn>=1.2
numpy>=1.24
